revert: undo all AEPS JpaRepository scan and mapping changes since cb923e5

+++ backend-java/src/main/java/com/rupiksha/aeps/entity/PayoutTransaction.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
-import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -27,7 +26,7 @@ public class PayoutTransaction {
     private String userId;
 
     @Column(nullable = false)
-    private BigDecimal amount;
+    private Double amount;
 
     @Column(nullable = false)
     private String beneficiaryName;
